finaliando para mandar pro pythonanywhere

+++ usuarios/signals.py
@@ -24,15 +24,19 @@ def create_user_profile(sender, request, user, **kwargs):
 
 @receiver(pre_save, sender=PerfilUsuario)
 def delete_old_image(sender, instance, **kwargs):
-    if not instance.pk:
-        return False
+    if not instance.pk:  # Se o objeto é novo, não há imagem antiga para deletar
+        return
 
     try:
+        # Recupera a instância do banco de dados antes da atualização
         old_image = PerfilUsuario.objects.get(pk=instance.pk).foto_perfil
     except PerfilUsuario.DoesNotExist:
-        return False
+        return
 
+    # Recupera a nova imagem que será salva
     new_image = instance.foto_perfil
-    if not old_image == new_image:
-        if os.path.isfile(old_image.path):
+
+    # Se a imagem foi alterada e o arquivo antigo existe, exclua-o
+    if old_image and old_image != new_image:
+        if old_image.path and os.path.isfile(old_image.path):
             os.remove(old_image.path)
